requirements.txt
google-genai>=0.2.0
google-cloud-storage>=2.15.0
